requirements.txt
pdfplumber
tqdm
regex
nltk
ipykernel
chromadb
nomic
python-dotenv
fastapi
uvicorn[standard]
python-dotenv
nomic
langchain
langchain-community
faiss-cpu
requests
pydantic
google-generativeai
fastapi

langchain-core
langchain-community
langchain-groq

faiss-cpu

requests
pydantic
